+++ src/config.rs
@@ -53,6 +53,9 @@ pub struct Metrics {
 
     #[serde(default)]
     pub populate_on_start: bool,
+
+    #[serde(default = "crate::config::default::stuck_packets")]
+    pub stuck_packets: bool,
 }
 
 mod default {
@@ -60,6 +63,10 @@ mod default {
 
     pub fn comet_version() -> CometVersion {
         CometVersion::V0_34
+    }
+
+    pub fn stuck_packets() -> bool {
+        true
     }
 }